Backend Api/requirements.txt
Flask==2.1.1
Flask-CORS==3.0.10
huggingface_hub==0.17.1
Werkzeug==2.2.3  
gunicorn==20.1.0
redis==5.0.0
python-dotenv
sentence-transformers
numpy
pandas
openpyxl
twilio==8.12.0